fix for rspec 3 support
also dropping 2x specific mods
