+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
 gem 'rake'
-gem 'rspec',    '>= 2.12'
+gem 'rspec',    '~> 3.1'
 gem 'minitest', '>= 4.3'
 gem 'sorcerer', '>= 0.3.7'
